Use makerepo function
Added to Merge Mergesort README

Added to the task and to the relevant commands

Added useful commands section to amend/README.

Added to the task in reorder-the-history/README.

Also added to useful commands.

Add clone step to README Gif

Added useful commands and more to reset/README

Added useful commands section and a further explanation section.

Added the task section to bad-commit/README.md

Added a detailed the task section and added to useful commands.

Added The task to commit-on-wrong-branch/README.md

Also added Useful commands

Added to The task in merge-conflict/README.md

Added detailed steps to The tasks and updated the Relevant commands

Added to The task in reverted-merge/README.md

Added detail to The task and added to Useful commands.

Added The task in save-my-commit/README.md

Updated the README to be more similar to others.
Made The task slightly clearer and added some hints (via googling)

Added to the task detached-head/README.

* Made The task more standard and added some detail.
* Added useful commands.