add git switch to exercise basic-branching
Cleaned up markdown in README files

Mostly fixing markdown-lint errors

Update to use git switch instead of git checkout

The new commands `git switch`and `git restore` were introduced
in 2.23 back in August 2019, and are much cleaner than the
old `git checkout`, so I believe that we
should be teaching the new way, not the old.

While `git switch` and `git restore` are in principal still marked as
experimental in the documentation, they are here to stay.
Even output from commands like git status,
have gone away from suggesting `git checkout`.

This commit goes through all the README files and makes appropriate
changes to recomend `git switch` instead of `git checkout`
(and `git switch -c` instead of `git checkout -b`)

The actual setup scripts are still left untouched for maximum compatablility.
If you have e.g. participants in a training using very old Git versions,
it is still a lot easier to tell them to use `git checkout`
when `git switch`is mentioned,
than it is to debug scripts that are failing miserably.

Update README.md

In the training the students learn multiple ways of restoring a file. Both checkout and reset is fine alternatives, but I think restore is the most accurate command for the job. Both checkout and reset also has other purposes where their use makes more sense. For instance reset is used whenever you want to move HEAD.

Add a comment about feature/uppercase

It is not obvious to beginners that `feature/uppercase` is a legal branch name and that this is what the exercise actually suggest,
so I added a comment to clarify this.

Improve wording in ignore exercise

Avoid confirmation prompt in mergesort setup script

Fixes #271